Move YamlUserConfiguration to authority module.
