+++ kernel/authority/core/src/main/java/org/apache/shardingsphere/authority/yaml/config/YamlUserConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.user.yaml.config;
+package org.apache.shardingsphere.authority.yaml.config;
 
 import lombok.Getter;
 import lombok.Setter;
