+++ kernel/authority/core/src/test/java/org/apache/shardingsphere/authority/yaml/swapper/YamlEngineUserConfigurationTest.java
@@ -15,15 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.user.yaml.swapper;
+package org.apache.shardingsphere.authority.yaml.swapper;
 
-import org.apache.shardingsphere.infra.metadata.user.yaml.config.YamlUserConfiguration;
+import org.apache.shardingsphere.authority.yaml.config.YamlUserConfiguration;
 import org.apache.shardingsphere.infra.util.yaml.YamlEngine;
+import org.hamcrest.CoreMatchers;
+import org.hamcrest.MatcherAssert;
 import org.junit.Test;
 
 import java.util.Properties;
 
-import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 public final class YamlEngineUserConfigurationTest {
@@ -31,19 +32,19 @@ public final class YamlEngineUserConfigurationTest {
     @Test
     public void assertUnmarshal() {
         YamlUserConfiguration actual = YamlEngine.unmarshal("password: pwd", YamlUserConfiguration.class);
-        assertThat(actual.getPassword(), is("pwd"));
+        assertThat(actual.getPassword(), CoreMatchers.is("pwd"));
     }
     
     @Test
     public void assertSecureUnmarshalProperties() {
         Properties actual = YamlEngine.unmarshal("password: pwd", Properties.class);
-        assertThat(actual.getProperty("password"), is("pwd"));
+        MatcherAssert.assertThat(actual.getProperty("password"), CoreMatchers.is("pwd"));
     }
     
     @Test
     public void assertMarshal() {
         YamlUserConfiguration actual = new YamlUserConfiguration();
         actual.setPassword("pwd");
-        assertThat(YamlEngine.marshal(actual), is("password: pwd" + System.lineSeparator()));
+        MatcherAssert.assertThat(YamlEngine.marshal(actual), CoreMatchers.is("password: pwd" + System.lineSeparator()));
     }
 }
